agents/merger_verifier.py
```python
from models.merge_verify import MergeVerify
from models.user_input import UserInput
from utils.openai_api import call_openai

class MergerVerifier:
    def __init__(self):
        self.merge_verify = MergeVerify()
        self.user_input = UserInput()

    def merge_parts(self, parts):
        merged_content = "\n".join(parts)
        self.merge_verify.set_merged_content(merged_content)
        return merged_content

    def verify_content(self, merged_content):
        user_inputs = self.user_input.get_user_inputs()
        subject = user_inputs.get('subject')
        keywords = user_inputs.get('keywords')
        target_content_type = user_inputs.get('target_content_type')
        word_count = user_inputs.get('word_count')

        # Call OpenAI API to verify the content
        verification_result = call_openai(merged_content, subject, keywords, target_content_type, word_count)
        self.merge_verify.set_verification_result(verification_result)
        return verification_result

    def merge_and_verify(self, parts):
        merged_content = self.merge_parts(parts)
        verification_result = self.verify_content(merged_content)
        return merged_content, verification_result
```