tests/test_merger_verifier.py
```python
import unittest
from agents.merger_verifier import MergerVerifier
from models.merge_verify import MergeVerifySchema
from models.user_input import UserInputSchema

class TestMergerVerifier(unittest.TestCase):

    def setUp(self):
        self.merger_verifier = MergerVerifier()
        self.user_input = UserInputSchema().load({
            'subject': 'AI',
            'keywords': ['Artificial Intelligence', 'Machine Learning'],
            'target_content_type': 'blog',
            'word_count': 1000
        })
        self.merge_verify = MergeVerifySchema().load({
            'project_id': '1',
            'outline_id': '1',
            'part_id': '1',
            'step_id': '1',
            'merge_verify_id': '1',
            'user_input_id': '1'
        })

    def test_merge_verify(self):
        result = self.merger_verifier.merge_verify(self.merge_verify, self.user_input)
        self.assertIsNotNone(result)
        self.assertEqual(result['merge_verify_id'], self.merge_verify['merge_verify_id'])

    def test_validate_content(self):
        content = "This is a test content for AI blog."
        result = self.merger_verifier.validate_content(content, self.user_input)
        self.assertTrue(result)

if __name__ == '__main__':
    unittest.main()
```